Reworked optimization.
Optimize (or formerly crop_to_dynamic_range) is now once again done
on init. Can be turned off with optmize parameter. This is to keep users
from manually optimizing after every model creation (which should be
done anyways). We are now smarter when to optimize, to
keep unnecessary warnings from popping up, which is the reason it
was removed from init in the first place.